requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic_core==2.6.3
pandas==2.1.0
SQLAlchemy==2.0.21
psycopg2-binary
python-dotenv==1.0.0
python-multipart==0.0.6
mail-parser==3.15.0
minio==7.1.16
boto3==1.28.52
pytz==2023.3.post1
sentry-sdk==1.31.0